% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bal.r
\name{bal}
\alias{bal}
\title{Bal: Approximate balancing weights}
\usage{
bal(Xt, X, objective = c("entropy", "l1", "l2"), tol = 0, solv = "MOSEK")
}
\arguments{
\item{Xt}{Target means (K vector)}

\item{X}{Data matrix (N × K)}

\item{objective}{Loss function for weights. One of entropy, L1, L2.}

\item{tol}{Tolerance for imbalance. 0 by default, which corresponds with exact balance.}

\item{solv}{Solver name to pass to CVXR solve}
}
\value{
n-vector of weights
}
\description{
Solve for a set of balancing weights subject to a convex loss.
}
