% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_bal.r
\name{bal_lbfgs}
\alias{bal_lbfgs}
\title{Balancing weights using dual formulation implemented in torch (autodiff via torch)}
\usage{
bal_lbfgs(
  X0,
  X1,
  lr = 0.01,
  num_iterations = 50,
  objective = c("entr", "l2"),
  noi = FALSE,
  debug = FALSE
)
}
\arguments{
\item{X0}{matrix of donor covariates}

\item{X1}{target moments}

\item{lr}{learning rate for optimizer}

\item{num_iterations}{number of iterations to run the optimizer}

\item{objective}{loss function (either "entr" or "l2")}

\item{noi}{boolean - print last value of gradient}

\item{debug}{print lots of intermediate data}
}
\value{
vector of weights
}
\description{
Balancing weights using dual formulation implemented in torch (autodiff via torch)
}
