% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pogs_bal.r
\name{bal_pogs}
\alias{bal_pogs}
\title{Balancing weights using pogs}
\usage{
bal_pogs(X0, X1, loss = c("entr", "l2"), tol = 0)
}
\arguments{
\item{X0}{matrix of donor covariates}

\item{X1}{target moments}

\item{loss}{one of "entr" or "l2"}

\item{tol}{tolerance for balance constraint violation}
}
\value{
vector of weights
}
\description{
Balancing weights using pogs
}
