% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebweights.R
\name{eb_solve_primal}
\alias{eb_solve_primal}
\title{Entropy balancing by solving primal
Mainly for didactic purposes; don't use.}
\usage{
eb_solve_primal(X1m, X0, solv = "MOSEK")
}
\arguments{
\item{X1m}{K-vector of target means}

\item{X0}{NXK matrix of covariates}

\item{solv}{Solver for CVXR ()}
}
\description{
Entropy balancing by solving primal
Mainly for didactic purposes; don't use.
}
