% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_bal.r
\name{ebal_torch}
\alias{ebal_torch}
\title{ebal implementation with autodiff via torch}
\usage{
ebal_torch(X0, X1)
}
\arguments{
\item{X0}{donor units matrix}

\item{X1}{target moments}
}
\value{
vector of weights
}
\description{
ebal implementation with autodiff via torch
}
