% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.r
\name{hainmueller}
\alias{hainmueller}
\title{DGP from Frolich (2007) / Hainmueller (2012)}
\usage{
hainmueller(error_design, outcome_design, n = 10000)
}
\arguments{
\item{error_design}{string (1, 2, or 3) for low, medium, or high overlap.}

\item{outcome_design}{string (1, 2, or 3) for linear, complex, or quadratic outcome model.}

\item{n}{sample size}
}
\description{
DGP from Frolich (2007) / Hainmueller (2012)
}
