% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrweights.R
\name{qr_solve_primal}
\alias{qr_solve_primal}
\title{Balancing weights to minimise quadratic divergence between uniform and solution weights. These are Renyi weights for α = 1}
\usage{
qr_solve_primal(X1m, X0, base_weights = NULL, solv = "MOSEK")
}
\arguments{
\item{solv}{solver for CVXR}

\item{X}{[data.frame, matrix] table of covariates for source sample (n x k)}

\item{target_moments}{[vector, matrix] target moments to match X columns (1 x k)}

\item{base}{weights n_0 vector of base weights}
}
\value{
list containing a n-vector of weights
}
\description{
Balancing weights to minimise quadratic divergence between uniform and solution weights. These are Renyi weights for α = 1
}
